+++ tools/release/app/apps/__init__.py
@@ -1,0 +1,3 @@
+from .release import release_app
+
+__all__ = ["release_app"]
