+++ tools/release/app/apps/release/version.py
@@ -2,7 +2,7 @@ import json
 import os
 from typing import Dict
 
-from toolbelt.constants import DATA_DIR
+from app.constants import DATA_DIR
 from datetime import datetime
