+++ tools/release/app/data/latest.json
