+++ tools/release/app/manager/__init__.py
@@ -1,0 +1,3 @@
+from .player_version_manager import PlayerVersionManager
+
+__all__ = ["PlayerVersionManager"]
