+++ tools/release/app/tools/esigner/esigner.py
@@ -1,8 +1,8 @@
 import shlex
 import subprocess
 
-from toolbelt.config import config
-from toolbelt.exceptions import EsignerError
+from app.config import config
+from app.exceptions import EsignerError
 
 
 class Esigner:
