+++ tools/release/tests/client/test_docker.py
@@ -1,6 +1,6 @@
 from tests.path import *
 from tests.testdata import read_file_as_json
-from toolbelt.client import DockerClient
+from app.client import DockerClient
 
 namespace = "planetariumhq"
 repo = "ninechronicles-headless"
