--- tools/release/toolbelt/apps/__init__.py
@@ -1,5 +1,0 @@
-from .k8s import k8s_app
-from .release import release_app
-from .update import update_app
-
-__all__ = ["k8s_app", "update_app", "release_app"]
