--- tools/release/toolbelt/client/__init__.py
@@ -1,6 +1,0 @@
-from .docker import DockerClient
-from .github import GithubClient
-from .session import BaseUrlSession
-from .slack import SlackClient
-
-__all__ = ["SlackClient", "GithubClient", "BaseUrlSession", "DockerClient"]
