--- tools/release/toolbelt/manager/__init__.py
@@ -1,4 +1,0 @@
-from .apv_history_manager import APVHistoryManager
-from .player_version_manager import PlayerVersionManager
-
-__all__ = ["APVHistoryManager", "PlayerVersionManager"]
